Add documentation to all Tile classes

+++ src/fi/tamk/cgeng/tiles/TileMap.java
@@ -3,23 +3,52 @@ package fi.tamk.cgeng.tiles;
 import java.awt.Graphics;
 import fi.tamk.cgeng.util.*;
 import fi.tamk.cgeng.engine.GameCam;
-
+/**
+ * TileMap is map of tiles. Map is used to draw
+ * tiles from TileSet to correct locations using
+ * Tiles id information.
+ */
 public class TileMap{
     
+    /**
+     * int array that contains id's of tiles. tileMap
+     * is used to draw tiles to correct locations.
+     */
     private int[][] tileMap;
+    /**
+     * TileSet is Set where Tiles information(texture, id
+     * width and height) is taken. 
+     */
     private TileSet tileSet;
 
-
+    /**
+     * Constructor that initializes tileMap and tileSet
+     * attributes.
+     * @param tileMap int Array of tile id's
+     * @param tileSet Set of tiles
+     */
     public TileMap(int[][] tileMap, TileSet tileSet){
         setTileSet(tileSet);
         setTileMap(tileMap);
     }
 
+    /**
+     * Constructor that loads tileMap from file and initializes
+     * tileSet
+     * @param path path of file
+     * @param width tileMaps width in tiles
+     * @param height tileMaps height in tiles
+     * @param tileSet tileSet where tiles info is taken. 
+     */
     public TileMap(String path, int width, int height, TileSet tileSet){
         loadMap(path, width, height);
         setTileSet(tileSet);
     }
 
+    /**
+     * Paints whole tileSet to screen.
+     * @param g map is rendered by Graphics class
+     */
     public void paint(Graphics g){
         for(int i = 0; i < tileMap.length; i++){
             for(int j = 0; j < tileMap[i].length; j++){
@@ -30,6 +59,13 @@ public class TileMap{
         }
     }
 
+    /**
+     * Paints whole tileSet to screen, developer is able to
+     * move all the tiles with xOffset and yOffset parameters.
+     * @param g map is rendered by Garphics class
+     * @param xOffset how much tiles are moved left
+     * @param yOffset how much tiles are moved to up
+     */
     public void paint(Graphics g, int xOffset, int yOffset){
         for(int i = 0; i < tileMap.length; i++){
             for(int j = 0; j < tileMap[i].length; j++){
@@ -40,6 +76,12 @@ public class TileMap{
         }
     }
 
+    /**
+     * Paints only tiles which given GameCam is looking at.
+     * Called by default in GameCam's paint method.
+     * @param g map is rendered by Garphics class
+     * @param cam GameCam that you are using
+     */
     public void paint(Graphics g, GameCam cam){
         for(int i = 0; i < tileMap.length; i++){
             for(int j = 0; j < tileMap[i].length; j++){
@@ -49,13 +91,20 @@ public class TileMap{
                    j*tileSet.getTileHeight() < cam.getY()+cam.getHeight()
                    ){
                     tileSet.getTile(tileMap[i][j])
-                    .paint(g,(i*tileSet.getTileWidth())-(int)cam.getX(),(j*tileSet.getTileHeight()-(int)cam.getY()),
+                    .paint(g,(i*tileSet.getTileWidth())-cam.getX(),
+                    (j*tileSet.getTileHeight()-cam.getY()),
                     tileSet.getTileWidth(),tileSet.getTileHeight());
                 }
             }
         }
     }
 
+    /**
+     * Loads tileMap from file.
+     * @param path path of file.
+     * @param width width of tileMap in tiles
+     * @param height height of tileMap in tiles
+     */
     private void loadMap(String path, int width, int height){
         String file = Utils.loadFileAsString(path);
         String[] tokens = file.split("\\s+");
@@ -68,14 +117,28 @@ public class TileMap{
         }
     }
 
+
+    /**
+     * Changes tileSet.
+     * @param tileSet new TileSet
+     */
     public void setTileSet(TileSet tileSet){
         this.tileSet = tileSet;
     }
 
+    /**
+     * Changes tileMap.
+     * @param tileMap new TileMap
+     */
     public void setTileMap(int[][] tileMap){
         this.tileMap = tileMap;
     }
 
+    /**
+     * Returns tile at tileMap's [x][y]
+     * @param x x coordinate of tileMap
+     * @param y y coordinate of tileMap
+     */
     public Tile tileAt(int x, int y){
         if(x > 0 && x < tileMap.length && y > 0 && y < tileMap[0].length){
             return tileSet.getTile(tileMap[x][y]);
@@ -83,6 +146,11 @@ public class TileMap{
         return tileSet.getTile(0);
     }
 
+    /**
+     * Returns tile at coordinate(in pixels).
+     * @param x x coordinate in pixels
+     * @param y y coordinate in pixels
+     */
     public Tile tileAtCoord(int x, int y){
         if(x/tileSet.getTileWidth() > 0 && 
             x/tileSet.getTileWidth() < tileMap.length &&
